01 Desember 2025 1. Edit Kode Models dan Services Product dan Customer 2. Membuat Fungsi Generate Kode Customer di Services 3. Membuat Fungsi Tambah Customer 4. Membuat Fungsi Edit Customer 5. Membuat Fungsi Generate Kode Product di Services 6. Membuat Fungsi Tambah Product 7. Membuat Fungsi Edit Product 8. Membuat Fungsi Delete Product

+++ lib/screens/cashier/cashier_pos_screen.dart
@@ -56,7 +56,7 @@ class _CashierPosScreenState extends State<CashierPosScreen> {
   int get totalHarga {
     int total = 0;
     for (var item in _cartItems) {
-      total += item.produk.hargaProduk * item.qty;
+      total += item.produk.hargaProduk.toInt() * item.qty;
     }
     return total;
   }
